+++ libs/ledger/include/ledger/chain/consensus/consensus_miner_interface.hpp
@@ -17,17 +17,18 @@
 //
 //------------------------------------------------------------------------------
 
-#include "ledger/chain/main_chain.hpp"
+#include <cstdint>
 
 namespace fetch {
 namespace ledger {
+
+class Block;
+
 namespace consensus {
 
 class ConsensusMinerInterface
 {
 public:
-  using Block = ledger::Block;
-
   ConsensusMinerInterface()          = default;
   virtual ~ConsensusMinerInterface() = default;
